// src/app/models/blast.model.ts
export interface BlastModel {
  id: number;
  blastText: string;
  lat: number;
  lng: number;
  userId: number;
  blastType: number;
  duration: number;
  distance: number;
  blastSavedCap: number;
  modifiedDate: string;
  userName: string;
  name: string;
  profileUrl: string;
  timeRemaining: number;
  comments: number;
  imageUri: string;
  saved: number;
  distanceAway: number;
  slides: string[];
}