// src/app/models/business.model.ts
export interface BusinessModel {
  id: number;
  name: string;
  address: string;
  address2: string;
  city: string;
  state: string;
  zip: string;
  phone: string;
  businessType: string;
  lat: number;
  lng: number;
  subscription: number;
  logoUrl: string;
  placesId: number;
  website: string;
}